refactor(role): remove deprecated role schemas and consolidate role handling in user schemas
